Replace audio implementation with improved version
- Add enhanced fetch_audio with resampling and normalization
- Support both bytes and path formats
- Automatic mono conversion
- Configurable sample rate and duration
- Better error handling and documentation

This replaces the previous basic implementation with a
production-ready version for VLM fine-tuning.
