+++ qwen-vl-utils/src/qwen_vl_utils/__init__.py
@@ -5,3 +5,16 @@ from .vision_process import (
     process_vision_info,
     smart_resize,
 )
+
+# Add audio utilities
+from .audio_process import fetch_audio
+
+# Export all functions
+__all__ = [
+    "extract_vision_info",
+    "fetch_image", 
+    "fetch_video",
+    "process_vision_info",
+    "smart_resize",
+    "fetch_audio",
+]
